% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tweets.R
\name{get_tweets}
\alias{get_tweets}
\title{get_tweets}
\usage{
get_tweets(user, startday = Sys.Date() - 30, stopday = Sys.Date(),
  verbose = FALSE)
}
\arguments{
\item{startday}{Earliest tweets to collect}

\item{stopday}{Latest tweets to collect}

\item{verbose}{whether the function will notify you if results are potentially problematic (i.e, private 
account, incorrect user ID, etc.).}
}
\value{
a tibble containing tweets
}
\description{
This function collects all of a user's tweets for a given timeframe (default is last 60 days).
}
\examples{
get_tweets("YouTube", startday= Sys.Date()-5, verbose=T)
}
