% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recip_followers.R
\name{recip_followers}
\alias{recip_followers}
\title{recip_followers}
\usage{
recip_followers(user, max.followers = 3e+05, max.friends = 20000,
  verbose = F)
}
\arguments{
\item{user}{a user screen name or ID}

\item{max.followers}{the maximum number of followers an account can have before it's considered a "large account" 
and not processed.}

\item{max.friends}{similar to max.followers, but for friends.}

\item{verbose}{whether the function should provide updates on its progress.}
}
\value{
a character vector of user IDs who are reciprocated followers with the user. If this vector has a length
of 0, then the user has no reciprocated followers
}
\description{
Finds the reciprocated followers of a given user. Handles rate limits and various problematic accounts, including
mistaken ID's, private accounts, suspended accounts, or accounts with too many friends or followers to process
in a timely way (this can be defined by the user).
}
