% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recip_followers_network.R
\name{recip_followers_network}
\alias{recip_followers_network}
\title{recip_followers_network}
\usage{
recip_followers_network(user, max.contacts = 500, verbose = F)
}
\arguments{
\item{user}{a user screen name or ID}

\item{max.contacts}{the maximum number of contacts (reciprocated followers) to process. If there are more than this
number, than a random sample of the contacts (n=max.contacts) will be used.}

\item{verbose}{whether the function should provide updates on its progress.}
}
\value{
user: the user ID of the original user

ncontacts: how many contacts the user has

funcstart: start time of function

funcstop: stop time of function

account.problems: information about potential problems with the account

sociogram: a matrix with identical rows and columns showing the connections among the user and their
contacts. The user's connection with themselves is marked as "self".
}
\description{
Maps out a user's 1.5 level ego network (who are their contacts, and whether these contacts 
know each other), based on reciprocated following. Repeatedly uses the recip_followers function to accomplish this.
Tie strength is binary: either two users are connected or they are unconnected.
}
