% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recip_mentioners.R
\name{recip_mentioners}
\alias{recip_mentioners}
\title{recip_mentioners}
\usage{
recip_mentioners(user, archive = NULL, preset.contacts = NULL,
  min.tie = 2, verbose = FALSE, trim.arch = T, sleepy = F, ...)
}
\arguments{
\item{user}{a user screen name or ID}

\item{archive}{pre-existing archive of tweets that can be used rather than having to re-collect tweets}

\item{preset.contacts}{vector of user screen names or IDs. If specified, all other users will be ignored in results.}

\item{min.tie}{the minimum tie strength (number of mentions exchanged) to include in the results.}

\item{verbose}{logical. How often the function should print updates on its progress.}

\item{trim.arch}{logical. Whether to remove a user's tweets from the archive if their tie strength with 
the original user is less than the minimum tie strength specified.}

\item{sleepy}{logical. If true, the function waits 60 seconds after processing each of the user's contacts}

\item{...}{additional arguments to pass to get_tweets within the function, i.e. to specify the timeframe}
}
\value{
mentions: a tibble that tracks the user's mentions of others. Each row specifies a user ID of a contact,
the mentions given to that contact, the mentions received from that contact, and the minimum of these last two
(the indicator of tie strength in this function)

archive: a tibble containing all tweets sent by the user and their contacts in the specified time frame
}
\description{
Tracks the mentions given and received for a user in a specified timeframe and archives the
tweets of the user and the people they talk to.
}
\details{
contacts whose accounts are private or suspended will have an NA in the mentions received column
of the mentions tibble

Depending on how sociable the user is, this function can run for a long time and collect
a lot of data, which can raise a variety of issues. Higher values of min.tie or specifying 
preset.contacts can prevent this. 
If trim.arch is set to FALSE, the archive can easily get quite 
large and overwhelm your system's memory.

For most users, the function has no problem with the rate limit. However, for users with a huge number
of contacts (>300), the function can break. Setting sleep to TRUE can prevent this.
}
