% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recip_mentioners_network.R
\name{recip_mentioners_network}
\alias{recip_mentioners_network}
\title{recip_mentioners_network}
\usage{
recip_mentioners_network(user, verbose = F, archive.to.csv = F,
  sleepy = F, ...)
}
\arguments{
\item{user}{a user screen name or ID}

\item{verbose}{logical. How often the function should print updates on its progress.}

\item{archive.to.csv}{logical. If set to TRUE, the archive will be written to a csv file
in the working directory rather than returned in the workspace.}

\item{sleepy}{logical. Used to avoid rate limit issues for users with huge number of contacts (>300).
If true, the function waits 60 seconds after processing each of the user's 
contacts in the first use of the recip_mentioners function (used throughout this function).
In subsequent uses, sleepiness isn't required because the tweets have already been collected.}

\item{...}{additional arguments to pass to recip_mentioners and get_tweets within the function.}
}
\value{
user: the user ID of the original user

ncontacts: how many contacts the user has

funcstart: start time of function

funcstop: stop time of function

account.problems: the output of the problem_account_check function for the original user

sociogram: a matrix with identical rows and columns showing the tie strength between the user
and their contacts. The user's connection with themselves is marked as NA.

archive: a tibble containing all tweets sent by the user and their contacts in the specified 
time frame.
}
\description{
Maps out a user's 1.5 level ego network (who are their contacts, and whether these contacts 
know each other), based on reciprocated mentions. Repeatedly uses the recip_mentioners function,
compiling an archive of all tweets sent by the user and their contacts in the timeframe. Tie strength
is defined as the minimum number of mentions exchanged between two users.
}
\details{
Note: this function ignores accounts that are private or suspended.
}
