% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_tweets.R
\name{reformat_tweets}
\alias{reformat_tweets}
\title{reformat_tweets}
\usage{
reformat_tweets(tweets)
}
\arguments{
\item{tweets}{a tibble of tweets}
}
\value{
a dataframe of tweets
}
\description{
Reformats the tibbles containing tweets from functions in either rtweet or twittnet and
makes them into a clean dataframe that can be written to a csv or other file
}
